apps/backend/requirements.txt
apache-airflow
requests
pandas
openmeteo_requests
psycopg2-binary
requests_cache
retry_requests
fastapi
uvicorn
sqlalchemy
python-dotenv==1.0.0
apache-airflow-providers-postgres
apache-airflow-providers-common-sql
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
typing_extensions
starlette